requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Image Processing
Pillow==10.1.0

# Data Processing and Reporting
pandas==2.1.3
openpyxl==3.1.2

# Environment Variables
python-dotenv==1.0.0

# Templating
jinja2==3.1.2

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Date/Time
python-dateutil==2.8.2
